Rewrite to `Int`
This replaces `Nat` with `Int` in the definitions of `Network` and
`Flow` (and everything that follows). With this, we avoid the annoying
proofs about subtracting natural numbers and backward flows.

A handful of places now have a non-negativity constraint, including the
final theorem - the matrix cannot contain negative numbers, because the
max flow is at least zero.